Logic pulled from DashboardPage.tsx to create a custom hook 'hooks/useDashboard' : On going bug issue NOT resolved
