+++ libs/web/feature-dashboard/src/lib/pages/DashboardPage.tsx
@@ -1,33 +1,21 @@
 import { PageLayout, Box } from '@daruma-board/web/design-system';
 import { GoalBoard } from '@daruma/web/feature-board';
-import { useRouter } from 'next/router';
-import { Header } from '../components';
+import { useDashboard } from '../hooks/useDashboard';
+import { Header } from '../components/Header';
 
-export function DashboardPage() {
-  // TODO: move to hook
-  const router = useRouter();
 
-  const onNewGoalClick = () => {
-    router.push('/board/new');
-  };
 
-  const onBoardClick = () => {
-    router.push('/board');
-  };
-
-  const onLogout = () => {
-    localStorage.removeItem('token');
-    localStorage.removeItem('localEmail');
-    router.push('/login');
-  };
-
+export const DashboardPage = () => {
+  const { onNewGoalClick, onBoardClick, onLogout} = useDashboard();
+  
+  
   return (
     <PageLayout>
       <Box
         sx={{
           outline: '1px solid red',
           borderRadius: '1rem',
-          backgroundColor: 'white', // TODO - replace with theme
+          backgroundColor: 'color.white', 
           width: '90vw',
           height: '70vh',
         }}
@@ -36,6 +24,7 @@ export function DashboardPage() {
           onNewGoalClick={onNewGoalClick}
           onBoardClick={onBoardClick}
           onLogout={onLogout}
+
         />
         <GoalBoard variant="private" />
       </Box>
